aula05/projeto02/requirements.txt
psycopg2-binary
mysql-connector-python
python-dotenv
pandas
matplotlib
pandas
matplotlib
